refactoring moved inner classes to outer
