+++ src/main/java/com/intellij/compiler/instrumentation/InstrumenterClassWriter.java
@@ -13,47 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.intellij.compiler.instrumentation;
 
 import org.objectweb.asm.ClassWriter;
 
 /**
-* @author Eugene Zhuravlev
-*         Date: 3/27/12
-*/
+ * @author Eugene Zhuravlev
+ *         Date: 3/27/12
+ */
 public class InstrumenterClassWriter extends ClassWriter {
-  private final InstrumentationClassFinder myFinder;
+    private final InstrumentationClassFinder myFinder;
 
-  public InstrumenterClassWriter(final int flags, final InstrumentationClassFinder finder) {
-    super(flags);
-    myFinder = finder;
-  }
+    public InstrumenterClassWriter(final int flags, final InstrumentationClassFinder finder) {
+        super(flags);
+        myFinder = finder;
+    }
 
-  protected String getCommonSuperClass(final String type1, final String type2) {
-    try {
-      final InstrumentationClassFinder.PseudoClass cls1 = myFinder.loadClass(type1);
-      final InstrumentationClassFinder.PseudoClass cls2 = myFinder.loadClass(type2);
-      if (cls1.isAssignableFrom(cls2)) {
-        return cls1.getName();
-      }
-      if (cls2.isAssignableFrom(cls1)) {
-        return cls2.getName();
-      }
-      if (cls1.isInterface() || cls2.isInterface()) {
-        return "java/lang/Object";
-      }
-      else {
-        InstrumentationClassFinder.PseudoClass c = cls1;
-        do {
-          c = c.getSuperClass();
+    protected String getCommonSuperClass(final String type1, final String type2) {
+        try {
+            final PseudoClass cls1 = myFinder.loadClass(type1);
+            final PseudoClass cls2 = myFinder.loadClass(type2);
+            if (cls1.isAssignableFrom(cls2)) {
+                return cls1.getName();
+            }
+            if (cls2.isAssignableFrom(cls1)) {
+                return cls2.getName();
+            }
+            if (cls1.isInterface() || cls2.isInterface()) {
+                return "java/lang/Object";
+            } else {
+                PseudoClass c = cls1;
+                do {
+                    c = c.getSuperClass();
+                }
+                while (!c.isAssignableFrom(cls2));
+                return c.getName();
+            }
+        } catch (final Exception e) {
+            throw new RuntimeException(e.toString(), e);
         }
-        while (!c.isAssignableFrom(cls2));
-        return c.getName();
-      }
     }
-    catch (final Exception e) {
-      throw new RuntimeException(e.toString(), e);
-    }
-  }
 }
